moved config to own package
