+++ pkg/config/config.go
@@ -1,0 +1,15 @@
+package config
+
+import (
+    "os"
+    "github.com/rs/cors"
+)
+
+func CorsOptions() cors.Options {
+    return cors.Options {
+        AllowedOrigins: []string{os.Getenv("CLIENT_ORIGIN_URL")},
+        AllowedMethods: []string{"GET"},
+        AllowedHeaders: []string{"Content-Type", "Authorization"},
+        MaxAge: 86400,
+    }
+}
